Shift useFlag hook and configcat library to common package, remove from studio
