+++ apps/studio/pages/project/[ref]/database/replication/[pipelineId].tsx
@@ -1,13 +1,12 @@
 import { useRouter } from 'next/router'
 import { useContext, useEffect } from 'react'
 
-import { FeatureFlagContext, useParams } from 'common'
+import { FeatureFlagContext, useFlag, useParams } from 'common'
 import { ReplicationPipelineStatus } from 'components/interfaces/Database/Replication/ReplicationPipelineStatus'
 import DatabaseLayout from 'components/layouts/DatabaseLayout/DatabaseLayout'
 import DefaultLayout from 'components/layouts/DefaultLayout'
 import { ScaffoldContainer, ScaffoldSection } from 'components/layouts/Scaffold'
 import { FormHeader } from 'components/ui/Forms/FormHeader'
-import { useFlag } from 'hooks/ui/useFlag'
 import { PipelineRequestStatusProvider } from 'state/replication-pipeline-request-status'
 import type { NextPageWithLayout } from 'types'
