+++ packages/common/feature-flags.tsx
@@ -141,3 +141,19 @@ export const FeatureFlagProvider = ({
 export const useFeatureFlags = () => {
   return useContext(FeatureFlagContext)
 }
+
+const isObjectEmpty = (obj: Object) => {
+  return Object.keys(obj).length === 0
+}
+
+export function useFlag<T = boolean>(name: string) {
+  const flagStore = useFeatureFlags()
+
+  const store = flagStore.configcat
+
+  if (!isObjectEmpty(store) && store[name] === undefined) {
+    console.error(`Flag key "${name}" does not exist in ConfigCat flag store`)
+    return false
+  }
+  return store[name] as T
+}
